+++ src/Plugin.php
@@ -92,11 +92,6 @@ class Plugin {
       return;
     }
 
-    // Blocks search indexing on search pages.
-    add_action('wp_head', __CLASS__ . '::wp_head');
-    // Disables Yoast adjacent links.
-    add_filter('wpseo_disable_adjacent_rel_links', __CLASS__ . '::wpseo_disable_adjacent_rel_links');
-
     // Removes coupon box from checkout.
     remove_action('woocommerce_before_checkout_form', 'woocommerce_checkout_coupon_form', 10);
 
@@ -141,11 +136,6 @@ class Plugin {
 
       // Ensure product details column is wide enough.
       add_filter('woocommerce_email_styles', __NAMESPACE__ . '\WooCommerce::woocommerce_email_styles');
-    }
-
-    if (get_option('_add_salutation_field') === 'yes') {
-      // Add salutation fields.
-      add_filter('woocommerce_checkout_fields', __NAMESPACE__ . '\WooCommerce::woocommerce_checkout_fields');
     }
 
     // Enqueues plugin scripts.
@@ -225,28 +215,6 @@ class Plugin {
       'instruction_placement' => 'label',
       'active' => 1,
     ]);
-  }
-
-  /**
-   * Blocks search indexing on search pages.
-   *
-   * @implements wp_head
-   */
-  public static function wp_head() {
-    $noindex_second_page = get_option(Plugin::L10N . '_robots_noindex_secondary_product_listings');
-
-    if (is_search() || (is_paged() && $noindex_second_page)) {
-      echo '<meta name="robots" content="noindex">';
-    }
-  }
-
-  /**
-   * Disables Yoast adjacent links.
-   *
-   * @implements wpseo_disable_adjacent_rel_links
-   */
-  public static function wpseo_disable_adjacent_rel_links() {
-    return get_option(Plugin::L10N . '_wpseo_disable_adjacent_rel_links');
   }
 
   /**
